Creating cutting widget for tile managment.

+++ App/src/Model/Managers/Strategies/ExtendStrategy.cs
@@ -18,7 +18,7 @@ namespace App.Model.Managers.Strategies
                 .Where(l => l < Selected.Left)
                 .OrderByDescending(t => t);
 
-            var left = candidates.Any() ? candidates.First() : (int?) null;
+            var left = candidates.Any() ? candidates.First() : (float?) null;
 
             var r = Selected;
             var rect = new Rect(left ?? r.Left, r.Top, r.Right, r.Bottom);
@@ -34,7 +34,7 @@ namespace App.Model.Managers.Strategies
                 .Where(l => l > Selected.Right)
                 .OrderBy(t => t);
 
-            var right = candidates.Any() ? candidates.First() : (int?) null;
+            var right = candidates.Any() ? candidates.First() : (float?) null;
 
             var r = Selected;
             var rect = new Rect(r.Left, r.Top, right ?? r.Right, r.Bottom);
@@ -50,7 +50,7 @@ namespace App.Model.Managers.Strategies
                 .Where(l => l < Selected.Top)
                 .OrderByDescending(t => t);
 
-            var top = candidates.Any() ? candidates.First() : (int?) null;
+            var top = candidates.Any() ? candidates.First() : (float?) null;
 
             var r = Selected;
             var rect = new Rect(r.Left, top ?? r.Top, r.Right, r.Bottom);
@@ -66,7 +66,7 @@ namespace App.Model.Managers.Strategies
                 .Where(l => l > Selected.Bottom)
                 .OrderBy(t => t);
 
-            var bottom = candidates.Any() ? candidates.First() : (int?) null;
+            var bottom = candidates.Any() ? candidates.First() : (float?) null;
 
             var r = Selected;
             var rect = new Rect(r.Left, r.Top, r.Right, bottom ?? r.Bottom);
